+++ 07. cli-runner-redo/cli-runner.go
@@ -3,7 +3,6 @@ package main
 import (
 	_ "flag"
 	_ "fmt"
-	cliHelper "github.com/nf/go-lang-testing/shared/cli-helper"
 	_ "io"
 	_ "io/ioutil"
 	_ "log"
@@ -13,17 +12,17 @@ import (
 )
 
 func main() {
-	writingConfig := cliHelper.ProcessConfigProperties{
-		Title:             "",
-		Description:       "",
-		ParentProcessName: "",
-		Id:                "",
-		ParentId:          "",
-		GlobalSessionId:   "",
-		Config:            cliHelper.WritingConfiguration{},
-	}
+	//writingConfig := cliHelper.ProcessConfigProperties{
+	//	Title:             "",
+	//	Description:       "",
+	//	ParentProcessName: "",
+	//	Id:                "",
+	//	ParentId:          "",
+	//	GlobalSessionId:   "",
+	//	Config:            cliHelper.WritingConfiguration{},
+	//}
 
-	processConfig := cliHelper.CreateDefaultLogWriterConfig(writingConfig)
-	processConfig.Writers.AttachErrorWriter()
+	//processConfig := cliHelper.CreateDefaultLogWriterConfig(writingConfig)
+	//processConfig.Writers.AttachErrorWriter()
 
 }
