+++ 09. cli-and-http-request/clihelper/AddLoggers.go
@@ -1,6 +1,6 @@
 package clihelper
 
-import . "github.com/nafrose/exploring/clirunner/clihelper/Structs"
+import . "github.com/nafrose/exploring/clirunner/clihelper/structs"
 
 type CliBindingPropertiesBuilder struct {
 	cliBindingProperties *CliBindingProperties
@@ -13,21 +13,18 @@ func NewCliBindingPropertiesBuilder() *CliBindingPropertiesBuilder {
 func (cli *CliBindingPropertiesBuilder) Build() *CliBindingProperties {
 	return cli.cliBindingProperties
 }
-
-func (cli *CliBindingPropertiesBuilder) WritersCollectionBuild() *WritersCollectionBuilder {
-	return &WritersCollectionBuilder{*it}
-}
-
-func (cli *CliBindingPropertiesBuilder) CmdRunningInfoBuild() *CmdRunningInfoBuilder {
-	return &CmdRunningInfoBuilder{*it}
-}
-
-func (cli *CliBindingPropertiesBuilder) CmdBuild() *CmdBuilder {
-	return &CmdBuilder{*it}
-}
-
-type LoggerBuilder struct {
-}
+//
+//func (cli *CliBindingPropertiesBuilder) WritersCollectionBuild() *WritersCollectionBuilder {
+//	return &WritersCollectionBuilder{*it}
+//}
+//
+//func (cli *CliBindingPropertiesBuilder) CmdRunningInfoBuild() *CmdRunningInfoBuilder {
+//	return &CmdRunningInfoBuilder{*it}
+//}
+//
+//func (cli *CliBindingPropertiesBuilder) CmdBuild() *CmdBuilder {
+//	return &CmdBuilder{*it}
+//}
 
 //func (o *OutputLoggers) Add(writer Writer) *OutputLoggers {
 //	result := make([]Writer, 0, 10)
