--- 09. cli-and-http-request/clihelper/Structs/WriterConfiguration.go
@@ -1,8 +1,0 @@
-package clihelper
-
-type WriterConfiguration struct {
-	IsJsonLoggingEnabled bool
-	IsHttpEnabled        bool
-	ShouldWriteToFile    bool
-	WriteToFilelocation  string
-}
