+++ 09. cli-and-http-request/clihelper/WriteUsingOutputLoggers.go
@@ -1,6 +1,6 @@
 package clihelper
 
-import . "github.com/nafrose/exploring/clirunner/clihelper/Structs"
+import . "github.com/nafrose/exploring/clirunner/clihelper/structs"
 
 func WriteUsingOutputLoggers(
 	cliBindingProperties CliBindingProperties,
