+++ 09. cli-and-http-request/clihelper/interfaces/Add.go
