+++ 09. cli-and-http-request/clihelper/interfaces/Writer.go
@@ -1,6 +1,6 @@
 package clihelper
 
-import . "github.com/nafrose/exploring/clirunner/clihelper/Structs"
+import . "github.com/nafrose/exploring/clirunner/clihelper/structs"
 
 type Writer interface {
 	Writer(cliBindingProperties CliBindingProperties, line string)
