+++ 09. cli-and-http-request/clihelper/structs/CliBindingProperties.go
