+++ 09. cli-and-http-request/clihelper/structs/CmdRunningInfo.go
