+++ 09. cli-and-http-request/clihelper/structs/StdInParameter.go
