+++ 09. cli-and-http-request/clihelper/structs/WritersCollection.go
@@ -1,6 +1,6 @@
 package clihelper
 
-import . "github.com/nafrose/exploring/clirunner/clihelper/Interfaces"
+import . "github.com/nafrose/exploring/clirunner/clihelper/interfaces"
 
 type WritersCollection struct {
 	OutputLoggers       []Writer
